Extract code to share
